+++ xtask/src/crosstool.rs
@@ -23,7 +23,7 @@ pub struct ConfigureCrosstool {
     /// Do not print cross log messages.
     #[clap(short, long)]
     pub quiet: bool,
-    /// Whether messages should use color output.
+    /// Coloring: auto, always, never
     #[clap(long)]
     pub color: Option<String>,
     /// The gcc version to configure for.