+++ xtask/src/target_info.rs
@@ -19,7 +19,7 @@ pub struct TargetInfo {
     /// Do not print cross log messages.
     #[clap(short, long)]
     pub quiet: bool,
-    /// Whether messages should use color output.
+    /// Coloring: auto, always, never
     #[clap(long)]
     pub color: Option<String>,
     /// Image registry.